build: phase 4 -- rn-iso builds to dist, dev/dist spawn paths, no daemon leak
tsdown builds four ESM+dts entries: cli (dist/cli.js, shebang), cache-manifest
(public export), and the two spawnable daemons supervisor-run and collector-run
as their own outputs. Deps (@rn-iso/metro, commander, chalk) and the
project-first @expo/fingerprint createRequire stay external.

spawn-entry.ts resolves the supervisor/collector script path in BOTH layouts:
the dev src/*.ts tree (import.meta.url contains /src/) and the built dist/*.js
sibling of dist/cli.js. Wired at all three spawn sites.

Daemon-leak hazard, this time the collector: android.ts imported readCollectors
from collector/run.ts, pulling run.ts's main()-guard into a chunk shared with
cli.js so its import.meta.url no longer matched. Fixed as supervisor/state.ts
was: new collector/state.ts holds the guard-free helpers; run.ts is the entry
only.

package.json: bin -> dist/cli.js, exports['./cache-manifest'] -> dist, files
[dist, skill, LICENSE, README]. Built CLI verified (--version, 11 commands,
status --json, both daemon entries fire main, cache-manifest exports). Dev CLI
+ typecheck 0 + vitest 52/1412 unaffected. Spawn-from-built against a real
project is deferred to e2e-native.
